move stuff into multiple files, and stop exporting the extension
