+++ packages/apollo-engine-reporting/tsconfig.json
@@ -8,7 +8,9 @@
     "outDir": "lib",
     "lib": ["es2017", "dom", "esnext.asynciterable"],
     "strict": true,
-    "strictPropertyInitialization": false
+    "strictPropertyInitialization": false,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true
   },
   "include" : [
     "src/**/*"
